\name{Dp}
\alias{Dp}
\title{Target dose, EDp}
\usage{
Dp(T, p)
}
\arguments{
\item{T}{A numeric vector. Model parameter values}

\item{p}{A numeric. Define EDp}
}
\description{
This finds the target dose level EDp which is the dose producing p percentage of the difference between the maximum and the minimum responses.
}
