
\name{DS1}
\alias{DS1}
\title{Sensitivity function of c-optimality criterion for the EDp}
\usage{
DS1(T, x, inv, p, order)
}
\arguments{
\item{T}{A numeric vector. Model parameter values for the 5PL model}

\item{x}{A numeric. A given dose level}

\item{inv}{A numeric matrix. Inverse infromation matrix for the 5PL model}

\item{p}{A numeric. Define EDp}

\item{order}{numeric. The number of model paraemters}
}
\description{
This evaluates the sensitivity function of the c-optimality criterion for the EDp at the given value of x.
}
