
\name{ds11}
\alias{ds11}
\title{Sensitivity function of Ds-optimality criterion}
\usage{
ds11(T, x, inv, inv1, order)
}
\arguments{
\item{T}{A numeric vector. Model parameter values for the 5PL model}

\item{x}{A numeric. A given dose level}

\item{inv}{A numeric matrix. Inverse information matrix for the 5PL model}

\item{inv1}{A numeric matrix. Inverse information matrix for the 4PL model}

\item{order}{numeric. The number of model paraemters}
}
\description{
This evaluates the sensitivity function of the Ds-optimality criterion at the given value of x.
}
