\name{f}
\alias{f}
\title{Gradient of the mean function}
\usage{
f(T, x, order)
}
\arguments{
\item{T}{A numeric vector. Model parameter values}

\item{x}{A numeric. A given dose level}

\item{order}{numeric. The number of model paraemters}
}
\description{
This gives the first derivative of the mean response function (3, 4, or 5PL) with respect to the model parameters.
}
