
\name{g}
\alias{g}
\title{Partial derivative of the EDp with respect to the model parameters}
\usage{
g(T, p)
}
\arguments{
\item{T}{A numeric vector. Model parameter values}

\item{p}{A numeric. Define EDp}
}
\description{
Partial derivative of the EDp with respect to the model parameters
}
