
\name{infor}
\alias{infor}
\title{Obtain a information matrix at a single design point}
\usage{
infor(T, x, order)
}
\arguments{
\item{T}{A numeric vector. Model parameter values}

\item{x}{numeric. A single design point(dose level)}

\item{order}{numeric. The number of model paraemters}
}
\description{
Obtain a information matrix at a single given design point x.
}
