\name{Inv}
\alias{Inv}

\title{Adjusting invere information matrix being not singular

}
\description{When the inverse matrix is a singular, it is adjusted to become nonsingular matrix.

}
\usage{
Inv(M, I)
}

\arguments{
  \item{M}{A numeric matrix

}
  \item{I}{Identity matrix

}
}
