\name{S_weight}
\alias{S_weight}
\title{Newton Raphson method to get optimal weights

}
\description{Newton Raphson method to get optimal weights for given design points "X".
}
\usage{
S_weight(X, T, e1, f, ...)
}

\arguments{
  \item{X}{A numeric vector. K dose levels for a given design

}
  \item{T}{A numeric vector. Model parameter values

}
  \item{e1}{Numeric. Threshold to stop the iteration.

}
  \item{f}{A character. D_weight for D-optimal weight; c_weight for c-optimal weight; DD_weight for Ds-optimal weight

}
  \item{\dots}{Default.

}
}
