
\name{upinfor}
\alias{upinfor}
\title{Obtain normalized Fisher information matrix}
\usage{
upinfor(W, T, X, order)
}
\arguments{
\item{W}{A numeric vector. The first K-1 weights for a given design}

\item{T}{A numeric vector. Model parameter values}

\item{X}{A numeric vector. K design points for a given design}

\item{order}{numeric. The number of model paraemters}
}
\description{
Obtain normalized Fisher information matrix for the 3, 4, 5PL models.
}
